Reorganize pe / pe_utils
cherry-picked from

- 7f1596d5b802567a0b81d95ddd0f6b3ebb234d6a: "Add support for CLI
  parsing"
- 94249fc2e2436e66c680446df4bac021929f3a45, "Other minor fixes in
  pe_utils.c"
- 04578d76555284d9cc8d0d902f408593e311fb7a, "Add pe_utils.c to Visual
  Studio 2015 project"
- 798d37a45822bc3c86dee67899b980c7a2870755, "Fix conditionals"
